requirements.txt
weaviate-client
langchain
transformers
torch
python-dotenv
